refactor(sync): move claude dynamic-sync into sandbox-core (close cloud→docker leak)
Phase 4 (dynamic). The cloud dynamic-config seed imported the manifest logic
and claude path helpers from @agentbox/sandbox-docker — a provider importing
another provider. Move the shared, provider-neutral pieces into the sync layer:

  - sync/agents/claude/paths.ts: encodeClaudeProjectsKey, BOX_CLAUDE_PROJECT_DIR,
    resolveClaudeMemoryDir (claude-specific host/box path helpers).
  - sync/concerns/dynamic.ts: the workflows+memory manifest logic
    (buildHostSyncManifest/computeSyncDelta/stageDynamicSyncTarball + types +
    BOX_WORKFLOWS_DIR/BOX_MEMORY_DIR/BOX_DYNAMIC_SYNC_MANIFEST), verbatim.

sandbox-docker's host-stage.ts and dynamic-sync.ts become thin re-export shims
so their existing importers (claude.ts, the docker index, the dynamic-sync
test) are untouched. cloud/dynamic-sync.ts now imports from @agentbox/sandbox-core
— leak gone. Docker create never consumed the manifest fns (it seeds
workflows/memory via the ~/.claude volume rsync), so this is cloud-runtime +
test only. seedDynamicConfig's cloud-specific exec/upload orchestration is
unchanged (its transport unification belongs with the Phase 7 driver).

Tests: docker dynamic-sync.test.ts (12) still green through the shims — a
cross-package guard that the moved surface is intact. Full gate green.
