+++ packages/sandbox-docker/src/host-stage.ts
@@ -180,47 +180,15 @@ const CLAUDE_RUNTIME_EXCLUDES = [
   'stats-cache.json',
 ];
 
-/**
- * Claude Code keys per-project state (memory, sessions, history) under
- * `~/.claude/projects/<encoded>/`, where `<encoded>` is the project's absolute
- * path with every non-alphanumeric char replaced by `-`. Inside every box the
- * workspace is `/workspace`, so its key is always `-workspace`. We duplicate
- * the rule here (rather than importing apps/cli's `encodeClaudeProjectsDir`)
- * because host-stage must not depend on the CLI package.
- */
-export function encodeClaudeProjectsKey(absPath: string): string {
-  return absPath.replace(/[^a-zA-Z0-9]/g, '-');
-}
-
-/** In-box Claude project dir for `/workspace` (fixed for every box). */
-export const BOX_CLAUDE_PROJECT_DIR = '/home/vscode/.claude/projects/-workspace';
-
-/**
- * Resolve the host's `~/.claude/projects/<encode(hostWorkspace)>/memory` dir,
- * or `null` when it's absent or empty (so callers no-op rather than seed an
- * empty tree). `hostHome` is overridable for tests.
- */
-export async function resolveClaudeMemoryDir(
-  hostWorkspace: string,
-  hostHome: string = homedir(),
-): Promise<string | null> {
-  if (hostWorkspace.length === 0) return null;
-  const memDir = join(
-    hostHome,
-    '.claude',
-    'projects',
-    encodeClaudeProjectsKey(hostWorkspace),
-    'memory',
-  );
-  if (!(await pathExists(memDir))) return null;
-  try {
-    const entries = await readdir(memDir);
-    if (entries.length === 0) return null;
-  } catch {
-    return null;
-  }
-  return memDir;
-}
+// Claude per-project path helpers moved to the shared sync layer
+// (`@agentbox/sandbox-core`) so the cloud dynamic-sync path reuses them without
+// importing this package. Re-exported here for existing importers (claude.ts,
+// the docker index, the dynamic-sync test).
+export {
+  encodeClaudeProjectsKey,
+  BOX_CLAUDE_PROJECT_DIR,
+  resolveClaudeMemoryDir,
+} from '@agentbox/sandbox-core';
 
 /**
  * Build the in-box `_claude.json` from the host's `~/.claude.json` (or a
